feat(prewrite): enable strict enforcement by default
